+++ ssg-point/app/(auth)/signup/auth/page.tsx
@@ -1,8 +1,11 @@
 'use client'
 import AuthBottom from "@/components/pages/auth/AuthBottom";
-import AuthTop from "@/components/pages/auth/AuthTop";
+import AuthHeaderBottom from "@/components/pages/auth/AuthHeaderBottom";
+import AuthTop from "@/components/pages/auth/AuthHeaderBottom";
+import AuthHeaderTop from "@/components/pages/auth/AuthHeaderTop";
 import CardAuthBody from "@/components/pages/auth/CardAuthBody";
 import PhoneAuthBody from "@/components/pages/auth/PhoneAuthBody";
+import { PhoneCardSelectBoxData } from "@/data/PhoneCardSelectBoxData";
 import { PhoneCardBoxType } from "@/types/PhoneCardBoxType";
 import { useState } from "react";
 
@@ -10,23 +13,11 @@ function Page(){
 
     const [selected,setSelected] = useState<string>("0");
 
-    const PhoneCardSelectBoxData: PhoneCardBoxType[] = [
-        {
-            id:"0",
-            title:"휴대폰인증",
-            url:"/assets/images/card&phoneIcon/phoneIcon.png",
-        },
-        {
-            id:"1",
-            title:"카드인증",
-            url:"/assets/images/card&phoneIcon/cardIcon.png",
-        }
-    ];
-
     return(
         <>
             <section className="pt-20">
-                <AuthTop
+                <AuthHeaderTop/>
+                <AuthHeaderBottom
                     typeList={PhoneCardSelectBoxData}
                     selected={selected}
                     setSelected={setSelected}
