--- ssg-point/app/(mypoint)/point-history/page.tsx
@@ -1,16 +1,0 @@
-import PointMainBottom from "@/components/pages/point/pointMain/PointMainBottom";
-import PointMainTop from "@/components/pages/point/pointMain/PointMainTop";
-import React from "react";
-
-function Page() {
-  return (  
-    <>
-    <section className="my-28">
-    <PointMainTop/>
-    <PointMainBottom/>
-    </section>
-    </>
-  );
-}
-
-export default Page;
