+++ ssg-point/components/pages/auth/AuthHeaderBottom.tsx
@@ -2,13 +2,11 @@ import React from "react";
 import Image from "next/image";
 import { PhoneCardBoxType } from "@/types/PhoneCardBoxType";
 
-const AuthTop = ( props: {
+const AuthHeaderBottom = ( props: {
     typeList: PhoneCardBoxType[];
     selected: string;
     setSelected: React.Dispatch<React.SetStateAction<string>>;
 }) => {
-
-const iconUrl : string = "/assets/images/signupIcon/";
 
 const {typeList, selected, setSelected } = props;
 
@@ -19,16 +17,8 @@ const handleClick = (e : any) => {
   
   return(
     <>
-      <div className="px-10 pb-5 pt-16 flex w-110">
-        <Image src={iconUrl + "checkIcon.png"} alt="" width={20} height={20} className="mr-2"/>
-        <Image src={iconUrl + "nonCheckicon.png"} alt="" width={20} height={20} className="mr-2"/>
-        <Image src={iconUrl + "nonCheckicon.png"} alt="" width={20} height={20} className="mr-2"/>
-        <Image src={iconUrl + "nonCheckicon.png"} alt="" width={20} height={20}/>
-      </div>
+      
       <div className="px-10">
-        <p className="text-2xl pb-1">
-          본인인증
-        </p>
         <b>
           <span className="text-[13px] text-[#767673]">
             본인인증 수단을 선택해주세요.
@@ -38,6 +28,7 @@ const handleClick = (e : any) => {
             {
                 typeList.map( item => (
                   <button className='box-border w-3/6 h-[88px] border flex justify-center items-center rounded-[6px] divide-[#e5e7eb]'
+                        key={item.id}
                         id={item.id}
                         onClick={(e) => handleClick(e)}
                         >
@@ -65,4 +56,4 @@ const handleClick = (e : any) => {
     );
 };
 
-export default AuthTop;
+export default AuthHeaderBottom;
