+++ ssg-point/data/PhoneCardSelectBoxData.ts
@@ -1,0 +1,14 @@
+import { PhoneCardBoxType } from "@/types/PhoneCardBoxType";
+
+export const PhoneCardSelectBoxData: PhoneCardBoxType[] = [
+  {
+      id:"0",
+      title:"휴대폰인증",
+      url:"/assets/images/card&phoneIcon/phoneIcon.png",
+  },
+  {
+      id:"1",
+      title:"카드인증",
+      url:"/assets/images/card&phoneIcon/cardIcon.png",
+  }
+];
